REFACTOR : js 코드 구조화
