+++ ShoppingList/index.html
@@ -33,5 +33,8 @@
     </main>
     <footer>화이팅 해야지</footer>
   </body>
-  <script type="module" src="/ShoppingList/index.js"></script>
+  <script type="module" src="/ShoppingList/js/bannerAni.js"></script>
+  <script type="module" src="/ShoppingList/js/cartManage.js"></script>
+  <script type="module" src="/ShoppingList/js/itemDisplay.js"></script>
+  <script type="module" src="/ShoppingList/js/navigation.js"></script>
 </html>
