+++ ShoppingList/js/navigation.js
@@ -1,0 +1,4 @@
+const homeBtn = document.querySelector(".homeBtn");
+homeBtn.addEventListener("click", () => {
+  window.location.href = "http://127.0.0.1:5500/ShoppingList/index.html";
+});
